+++ Ladybug/Model/Radar.swift
@@ -8,12 +8,19 @@
 
 import Foundation
 
-class Radar: Codable {
+public class Radar: Codable {
     let id: RadarID
     var metadata: RadarMetadata?
     let firstViewedDate: Date
     var lastViewedDate: Date
-    var favoritedDate: Date? = nil
+    var bookmarkedDate: Date? = nil
+
+    enum CodingKeys: String, CodingKey {
+        case id
+        case firstViewedDate
+        case lastViewedDate
+        case bookmarkedDate
+    }
 
     init(id: RadarID, metadata: RadarMetadata? = nil) {
         self.id = id
@@ -24,5 +31,25 @@ class Radar: Codable {
 
     var idString: String {
         return String(id.id)
+    }
+}
+
+extension Radar {
+    convenience init?(metadata: RadarMetadata) {
+        if let radarId = RadarID(string: metadata.number) {
+            self.init(id: radarId, metadata: metadata)
+        } else {
+            return nil
+        }
+    }
+}
+
+extension Radar {
+    var cellTitle: String {
+        return idString
+    }
+
+    var cellSubtitle: String {
+        return metadata?.title ?? "(No record on Open Radar)".localized()
     }
 }