+++ Ladybug/Model/RadarCollection.swift
@@ -18,6 +18,7 @@ class RadarCollection {
     var delegates: MulticastDelegate<RadarCollectionDelegate> = MulticastDelegate()
 
     static private let key = "com.elaborapp.Ladybug.RadarCollection"
+    static private let metadataKey = "com.elaborapp.Ladybug.RadarCollection.metadata"
 
     private var radars: [RadarID: Radar] = [:] {
         didSet {
@@ -34,9 +35,29 @@ class RadarCollection {
     }
 
     public func unarchive() {
+        let radarMetdata: [RadarID: RadarMetadata] = {
+            if let metadataData = UserDefaults.standard.object(forKey: RadarCollection.metadataKey) as? Data {
+                do {
+                    let metadata = try PropertyListDecoder().decode([RadarID: RadarMetadata].self, from: metadataData)
+                    print("Unarchived \(metadata.count) radar metadata")
+                    return metadata
+                } catch {
+                    print(error.localizedDescription)
+                }
+            }
+            return [:]
+        }()
+
         if let data = UserDefaults.standard.object(forKey: RadarCollection.key) as? Data {
             do {
                 let radars = try PropertyListDecoder().decode([RadarID: Radar].self, from: data)
+
+                for radar in radars {
+                    if let metadata = radarMetdata[radar.key] {
+                        radar.value.metadata = metadata
+                    }
+                }
+
                 self.radars = radars
                 print("Unarchived \(radars.count) radars")
             } catch {
@@ -47,9 +68,20 @@ class RadarCollection {
 
     public func archive() {
         do {
-            let data = try PropertyListEncoder().encode(radars)
-            UserDefaults.standard.set(data, forKey: RadarCollection.key)
-            print("Archived \(radars.count) radars")
+            let radarData = try PropertyListEncoder().encode(radars)
+            UserDefaults.standard.set(radarData, forKey: RadarCollection.key)
+
+            var radarMetdata: [RadarID: RadarMetadata] = [:]
+            for radar in radars {
+                if let metadata = radar.value.metadata {
+                    radarMetdata[radar.key] = metadata
+                }
+            }
+
+            let radarMetadataData = try PropertyListEncoder().encode(radarMetdata)
+            UserDefaults.standard.set(radarMetadataData, forKey: RadarCollection.metadataKey)
+
+            print("Archived \(radars.count) radars, \(radarMetdata.count) metadata")
         } catch {
             print(error.localizedDescription)
         }
@@ -59,8 +91,8 @@ class RadarCollection {
 
     public func upsert(radar: Radar) {
         if let existingRadar = radars[radar.id] {
-            if existingRadar.favoritedDate == nil {
-                existingRadar.favoritedDate = radar.favoritedDate
+            if existingRadar.bookmarkedDate == nil {
+                existingRadar.bookmarkedDate = radar.bookmarkedDate
             }
 
             if radar.lastViewedDate > existingRadar.lastViewedDate {
@@ -88,26 +120,35 @@ class RadarCollection {
         }
     }
 
-     /// Use when user toggle favorite for a radar
+     /// Use when user toggle bookmark for a radar
 
-    public func toggleFavorite(radarID: RadarID) throws {
+    public func toggleBookmark(radarID: RadarID) throws {
         if let existingRadar = radars[radarID] {
-            existingRadar.favoritedDate = existingRadar.favoritedDate != nil ? nil : Date()
+            existingRadar.bookmarkedDate = existingRadar.bookmarkedDate != nil ? nil : Date()
             notifyDidUpdate()
         }
     }
 
+    public func bookmark(radarIDs: [RadarID]) {
+        for radarID in radarIDs {
+            if let existingRadar = radars[radarID] {
+                existingRadar.bookmarkedDate = existingRadar.bookmarkedDate != nil ? existingRadar.bookmarkedDate : Date()
+            }
+        }
+        notifyDidUpdate()
+    }
+
     public func history() -> [Radar] {
         let radars = self.radars.values.sorted { (lhs, rhs) -> Bool in
-            return lhs.lastViewedDate > rhs.lastViewedDate
+            return lhs.firstViewedDate > rhs.firstViewedDate
         }
 
         return radars
     }
 
     public func bookmarks() -> [Radar] {
-        let radars = self.radars.values.filter { $0.favoritedDate != nil }.sorted { (lhs, rhs) -> Bool in
-            return lhs.favoritedDate! > rhs.favoritedDate!
+        let radars = self.radars.values.filter { $0.bookmarkedDate != nil }.sorted { (lhs, rhs) -> Bool in
+            return lhs.bookmarkedDate! > rhs.bookmarkedDate!
         }
 
         return radars